A disassembly of MON-1B with comments, function names etc.
This is an attempt at disassembling the MON-1B. It is by no means complete, comprehensive or accurate. But I try my best :)

Anyone interested in the details of how MON-1 works internally can look here and get a headstart on creating their own MONitor or working more closely with MON-1.